requirements.txt
beautifulsoup4==4.12.3
datasets==3.2.0
pandas==2.2.0
psycopg==3.2.4
python-dotenv==1.0.1
python-Levenshtein==0.23.0
streamlit==1.42.1
GitPython==3.1.37